requirements.txt
anyio==4.4.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
python-telegram-bot==21.5
pytz==2024.1
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
tornado==6.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
